+++ include/galois/filters/linear.h
@@ -10,8 +10,8 @@ namespace gs {
     private:
         SP_Signal<T> in_signal = nullptr;
         SP_Signal<T> out_signal = nullptr;
-        int in_size = 0;
-        int out_size = 0;
+        size_t in_size = 0;
+        size_t out_size = 0;
 
         SP_NArray<T> w = nullptr;
         SP_NArray<T> b = nullptr;
@@ -22,13 +22,13 @@ namespace gs {
         Linear(const bool for_clone_or_share) {}
         Linear(const Linear&) = delete;
         Linear& operator=(const Linear&) = delete;
-        Linear(int in_size, int out_size);
+        Linear(size_t in_size, size_t out_size);
 
         SP_Filter<T> share() override;
         SP_Filter<T> clone() override;
 
         void install_signals(const vector<SP_Signal<T>> &in_signals, const vector<SP_Signal<T>> &out_signals) override;
-        void set_dims(int batch_size) override;
+        void set_dims(size_t batch_size) override;
         void reopaque() override;
 
         vector<SP_NArray<T>> get_params() override;
