+++ include/galois/models/seq_encoder_decoder.h
@@ -16,25 +16,26 @@ namespace gs
         static default_random_engine galois_rn_generator;
 
     protected:
-        int max_len_encoder;
-        int max_len_decoder;
-        int input_size;
-        int output_size;
-        vector<int> hidden_sizes;
+        size_t max_len_encoder;
+        size_t max_len_decoder;
+        size_t input_size;
+        size_t output_size;
+        vector<size_t> hidden_sizes;
 
-        int train_seq_count = 0;
+        size_t train_seq_count = 0;
         SP_NArray<T> train_X = nullptr;
         SP_NArray<T> train_Y = nullptr;
-        int test_seq_count = 0;
+        size_t test_seq_count = 0;
         SP_NArray<T> test_X = nullptr;
         SP_NArray<T> test_Y = nullptr;
     public:
-        SeqEncoderDecoder(int max_len_encoder,
-            int max_len_decoder,
-            int input_size,
-            int output_size,
-            initializer_list<int> hidden_sizes,
-            int batch_size,
+        SeqEncoderDecoder(
+            size_t max_len_encoder,
+            size_t max_len_decoder,
+            size_t input_size,
+            size_t output_size,
+            initializer_list<size_t> hidden_sizes,
+            size_t batch_size,
             int num_epoch,
             T learning_rate,
             string optimizer_name);
