+++ include/galois/narray.h
@@ -25,47 +25,46 @@ namespace gs
     public:
         static default_random_engine galois_rn_generator;
 
-        explicit NArray(int m);
-        explicit NArray(int m, int n);
-        explicit NArray(int m, int n, int o);
-        explicit NArray(int m, int n, int o, int k);
-        explicit NArray(vector<int>);
+        explicit NArray(size_t m);
+        explicit NArray(size_t m, size_t n);
+        explicit NArray(size_t m, size_t n, size_t o);
+        explicit NArray(size_t m, size_t n, size_t o, size_t k);
+        explicit NArray(vector<size_t>);
         NArray() = delete;
         NArray(const NArray& other) = delete;
         NArray& operator=(const NArray&) = delete;
         ~NArray();
 
-        vector<int> get_dims() { return dims; }
-        int get_size() { return size; }
+        vector<size_t> get_dims() { return dims; }
+        size_t get_size() { return size; }
         T* get_data() { CHECK(data, "data should be non-empty"); return data; }
         bool opaque() { return data_opaque; }
         void reopaque() { data_opaque = true; }
         void setclear() { data_opaque = false; }
 
-//        void copy_from(const vector<int> &, const T*);
         void copy_from(const SP_NArray<T>);
-        void copy_from(const vector<int> &, const SP_NArray<T>);
-        void copy_from(const vector<int> &, int, const SP_NArray<T>);
-        void copy_from(const int, const int, const SP_NArray<T>);
+        void copy_from(const vector<size_t> &, const SP_NArray<T>);
+        void copy_from(const vector<size_t> &, size_t, const SP_NArray<T>);
+        void copy_from(const size_t, const size_t, const SP_NArray<T>);
         void uniform(T lower, T upper) {
             // future : move random generator to a single file
             uniform_real_distribution<T> distribution(lower, upper);
-            for (int i = 0; i < get_size(); i++) {
+            for (size_t i = 0; i < get_size(); i++) {
                 data[i] = distribution(galois_rn_generator);
             }
             data_opaque = false;
         }
         void normalize_for(int dim);
         void fill(T x) {
-            for (int i = 0; i < get_size(); i++) {
+            for (size_t i = 0; i < get_size(); i++) {
                 data[i] = x;
             }
             data_opaque = false;
         }
 
     private:
-        const vector<int> dims = {};
-        int size = 0;
+        const vector<size_t> dims = {};
+        size_t size = 0;
         T *data = nullptr;
         bool data_opaque = true;
     };
