+++ src/models/ordered_model.cc
@@ -198,8 +198,8 @@ namespace gs
     template<typename T>
     T OrderedModel<T>::train_one_batch(const bool update) {
         uniform_int_distribution<> distribution(0, train_count-1);
-        vector<int> batch_ids(batch_size);
-        for (int i = 0; i < batch_size; i++) {
+        vector<size_t> batch_ids(batch_size);
+        for (size_t i = 0; i < batch_size; i++) {
             batch_ids[i] = distribution(galois_rn_generator);
         }
 
@@ -235,9 +235,9 @@ namespace gs
     double OrderedModel<T>::test() {
         double correctness = 0;
 
-        for (int i = 0; i < test_count; i += batch_size) {
-            vector<int> batch_ids(batch_size);
-            for (int j = 0; j < batch_size; j++) {
+        for (size_t i = 0; i < test_count; i += batch_size) {
+            vector<size_t> batch_ids(batch_size);
+            for (size_t j = 0; j < batch_size; j++) {
                 batch_ids[j] = i+j;
             }
 
@@ -283,7 +283,7 @@ namespace gs
             auto start = chrono::system_clock::now();
 
             T loss = 0;
-            for (int i = 0; i < train_count/batch_size; i++) {
+            for (size_t i = 0; i < train_count/batch_size; i++) {
                 loss += train_one_batch();
             }
             loss /= T(train_count/batch_size);
